Extract methods from RDS Class (#1591)
