+++ apps/antalmanac/src/backend/routers/notifications.ts
@@ -1,7 +1,7 @@
 import { db } from '@packages/db/src/index';
 import { z } from 'zod';
 
-import { RDS } from '../../backend/lib/rds';
+import { RDS } from '../../backend/lib/rds/rds';
 import { procedure, router } from '../trpc';
 
 const NotifyOnSchema = z.object({
